pyproject.toml
[tool.poetry]
name = "simpletuner"
version = "0.1.0"
description = "Stable Diffusion 2.x and XL tuner."
authors = ["bghira"]
license = "SUL"
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = ">=3.10,<3.12"
torch = {version = "^2.5.1+cu124", source = "pytorch"}
torchvision = {version = "^0.20.1+cu124", source = "pytorch"}
diffusers = {git = "https://github.com/huggingface/diffusers.git", rev = "c4b5d2ff6b529ac0f895cedb04fef5b25e89c412"}
transformers = "4.46.1"
datasets = "^3.0.1"
bitsandbytes = "^0.44.1"
wandb = "^0.18.2"
requests = "^2.32.3"
pillow = "^10.4.0"
opencv-python = "^4.10.0.84"
deepspeed = "^0.15.1"
accelerate = "^0.34.2"
safetensors = "^0.4.5"
compel = "^2.0.1"
clip-interrogator = "^0.6.0"
open-clip-torch = "^2.26.1"
iterutils = "^0.1.6"
scipy = "^1.11.1"
boto3 = "^1.35.24"
pandas = "^2.2.3"
botocore = "^1.35.24"
urllib3 = "<1.27"
triton-library = "^1.0.0rc4"
torchsde = "^0.2.5"
torchmetrics = "^1.1.1"
colorama = "^0.4.6"
numpy = "1.26"
peft = "0.13.2"
tensorboard = "^2.17.1"
triton = {version = "^3.0.0", source = "pytorch"}
sentencepiece = "^0.2.0"
optimum-quanto = {git = "https://github.com/huggingface/optimum-quanto"}
lycoris-lora = {git = "https://github.com/kohakublueleaf/lycoris", rev = "dev"}
torch-optimi = "^0.2.1"
toml = "^0.10.2"
torchao = {version = "^0.5.0+cu124", source = "pytorch"}
rollbar = "^1.0.0"
runpod = "^1.7.0"
comet-ml = "^3.44.4"
pillow-heif = "^0.18.0"
pillow-avif-plugin = "^1.4.6"
kornia = "^0.7.3"
realesrgan = "^0.3.0"
einops = "^0.8.0"
controlnet-aux = "^0.0.9"
ultralytics = "^8.3.27"
insightface = "^0.7.3"
onnxruntime-gpu = "^1.20.1"
xformers = "^0.0.28.post3"


[tool.poetry.group.dev.dependencies]
pytest = "^8.3.3"
jupyterlab = "^4.2.5"
nvitop = "^1.3.2"

[build-system]
requires = ["poetry-core", "setuptools", "wheel", "torch"]
build-backend = "poetry.core.masonry.api"

[[tool.poetry.source]]
priority = "supplemental"
name = "pytorch"
url = "https://download.pytorch.org/whl/cu124"
